Refactor div/mod abstraction into a dedicated lemma catalogue
Split the 624-line DivModEncoding into three modules, following the
EVM.Expr simplifier idiom (pattern-directed rules + per-constructor
emit), so the lemmas live in one reviewable file:

- EVM.SMT.AbstractBase: shared SMT primitives, signed helpers, the
  div/mod taxonomy, term collectors/matchers, and `saturate` (the
  abstract-term closure that was tangled in mulEncoding's `let`).
- EVM.SMT.AbstractLemmas: the lemma catalogue and review surface. A
  `LemmaInst` sum type, `collectLemmas` (triggers), and `emitLemma`
  (SMT emission + soundness notes), one banner per lemma. GHC
  exhaustiveness now forces every constructor to have an emitter,
  replacing the old silent positional concat.
- EVM.SMT.DivModEncoding: slimmed to orchestration (mulEncoding) plus
  the div/mod ground-truth encoding.

Behavior-preserving: emitter bodies are verbatim and the emission order
is unchanged, so generated queries are identical. Public API unchanged.

Validated: lib + exe build clean (GHC 9.6.6); seven representative lemma
properties (const-cancel, scaled const-cancel, frac-reduce, nested-div,
mul-mono, mul-div-bound, telescope) prove under --abstract-arith, and a
false identity correctly fails, so soundness is preserved.
